refactor(package): move scoreType to shot.type
- move ShotScore class to domain.shot
- move Score class to domain.shot
- move scoreType package to shot.type
issue [#152(comment)](https://github.com/next-step/java-bowling/pull/151#discussion_r4177739030
