+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,6 +1,6 @@
 package bowling.domain.frame;
 
-import bowling.domain.ShotScore;
+import bowling.domain.shot.ShotScore;
 import bowling.domain.frame.score.FrameScore;
 
 import java.util.List;
