+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -1,6 +1,6 @@
 package bowling.domain.frame;
 
-import bowling.domain.ShotScore;
+import bowling.domain.shot.ShotScore;
 import bowling.domain.ShotScores;
 import bowling.domain.frame.score.DefaultFrameScore;
 import bowling.domain.frame.score.FrameScore;
