+++ src/test/java/bowling/domain/frame/NormalFrameTest.java
@@ -1,6 +1,6 @@
 package bowling.domain.frame;
 
-import bowling.domain.scoreType.ScoreType;
+import bowling.domain.shot.type.ScoreType;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
