+++ src/test/java/bowling/domain/frame/score/DefaultFrameScoreTest.java
@@ -1,6 +1,6 @@
 package bowling.domain.frame.score;
 
-import bowling.domain.scoreType.ScoreType;
+import bowling.domain.shot.type.ScoreType;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
