+++ src/test/java/bowling/domain/shot/ScoreTest.java
@@ -1,5 +1,6 @@
-package bowling.domain;
+package bowling.domain.shot;
 
+import bowling.domain.shot.Score;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
