+++ src/test/java/bowling/domain/shot/type/GutterCheckerTest.java
@@ -1,6 +1,6 @@
-package bowling.domain.scoreType;
+package bowling.domain.shot.type;
 
-import bowling.domain.Score;
+import bowling.domain.shot.Score;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
